Completed the new verification system
