+++ src/CmdTwitch/getPoints.ts
@@ -9,15 +9,6 @@ const getPointsFunc : twitchCmdType ={
         if(!data.user['user-id']) return;
         if(data.authUsers[data.user['user-id']] == "-1") return data.client.say(data.channel,`@${data.user.username}, your account is not linked yet. Do that first then try again.`);
         let discordUserID = data.authUsers[data.user['user-id']];
-        if(!discordUserID) {
-            // Get the user ID since it's not available when the stream is offline
-            const userData = await prisma.twitch.findFirst({
-                where: {
-                    twitchid: data.user['user-id']
-                }
-            })
-            if(userData) discordUserID = userData.memberid
-        }
         // Fetch the points
         const econData = await prisma.economy.findUnique({
             where: {
